Update sharem_wrapper.py
Updated the wrapper for the integration to make it a little easier to handle the data. The files are being opened up and converted into json, the file is then closed, and the json object is returned. This helps to streamline the collection of data in a few less lines of code in the parser which can be handled in the Sharem package.