fixed category issues; truncated cmd and powershell
Some files were not being put in the correct category, this is fixed now. Any paths that include exes/dlls are now both in the exe and the path categories.

Truncated any files names that are something like )cmd.exe to just cmd.exe. This also works for powershell.

Removed the [dot]com filetype to be found.