Update sharem_main.py
Update to include a reset globals, this fixes the issue with IndexError being thrown when multiple files are ran back to back. This was the only way to get it working. An issue with using globals besides having an increased issue having the ability to debug, is that you are not able to use the importlib reload function. The reload function isn't able to reset globals and after testing it, absolutely does not reset them. This new function allows for resetting all of the globals back to the original state before running a new file.